PR fixes
https://coveord.atlassian.net/browse/KIT-394
